requirements.txt
asgiref==3.9.1
coverage==7.10.7
Django==5.2.5
django-debug-toolbar==6.0.0
django-extensions==4.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
inflection==0.5.1
pydotplus==2.0.2
PyJWT==2.10.1
pyparsing==3.2.5
PyYAML==6.0.3
sqlparse==0.5.3
uritemplate==4.2.0
